requirements.txt
requests
python-dotenv
python-chess
stockfish